+++ mozjs-sys/src/jsglue.cpp
